+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unrelease] -
 
+- Fix Escalate timeline button execute RuleTicket
+
 
 ## [2.9.13] - 2025-03-31